requirements.txt
drms
astropy
numpy
pytorch
